% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expeer.R
\name{get_power}
\alias{get_power}
\alias{get_stratified_units}
\alias{get_randomized_units}
\title{Set up a randomized experiment}
\usage{
get_power(design, pop_filt)

get_stratified_units(design, pop_filt)

get_randomized_units(design, pop_filt, num_randomized_units)
}
\arguments{
\item{design}{List corresponding to an experimental design.}

\item{pop_filt}{\code{data.frame} containing one row per unit available for the
experiment.}

\item{num_randomized_units}{Number of units to randomize in the experiment.}
}
\value{
\code{get_power()} returns a \code{data.table} of power calculations, with one
row per metric. \code{get_stratified_units()} returns a \code{data.table} with one
row per unit and with columns for stratification. \code{get_randomized_units()}
returns a \code{data.table} with one row per unit and with columns for
stratification and random assignment to experiment arms.
}
\description{
\code{get_stratified_units()} is used by \code{get_randomized_units()} and does not
normally need to be called directly.
}
\examples{
library('data.table')

set.seed(2603)
n = 2000

pop = data.table(
  group_id = 1:n,
  state = rep_len(c('Karnataka', 'Haryana', 'Bihar'), n),
  num_members = rbinom(n, 50, 0.5),
  mean_engagement_1wk = rexp(n, 0.1),
  mean_engagement_2wk = rnorm(n, 50, 15),
  is_happy_2wk = rbinom(n, 1, 0.4))

pop[, district := paste(state, 'District', rep_len(1:2, n))]
pop[seq(5, n, 80), district := NA]
pop[seq(5, n, 41), num_members := NA]

########################################

design = yaml::read_yaml(
  system.file('extdata', 'expeer_design_01.yml', package = 'expeer'))

pop_filt = get_filtered_units(design, pop)

power_calcs = get_power(design, pop_filt)

num_randomized_units =
  power_calcs[metric_name == 'mean_engagement_2wk']$num_units_for_target_power

pop_rand = get_randomized_units(design, pop_filt, num_randomized_units)
}
\seealso{
\code{\link[=get_filtered_units]{get_filtered_units()}}
}
